refactor(mesh): set background color on body rather than content
- remove unnecessary Content component
- fix breaking layout in social page
- remove media page main bg-color
